backend/app/requirements.txt
Flask
flask-peewee
psycopg2-binary
flasgger
